--- app/src/main/java/com/jayanthr/spendisense/sms/SmsReceiver.kt
@@ -1,21 +1,0 @@
-package com.jayanthr.spendisense.sms
-
-import android.content.BroadcastReceiver
-import android.content.Context
-import android.content.Intent
-import android.provider.Telephony
-import android.util.Log
-
-class SmsReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context, intent: Intent) {
-        // Check if the received intent is for an SMS
-        if (intent.action == Telephony.Sms.Intents.SMS_RECEIVED_ACTION) {
-            val messages = Telephony.Sms.Intents.getMessagesFromIntent(intent)
-            for (message in messages) {
-                val sender = message.displayOriginatingAddress
-                val content = message.displayMessageBody
-                Log.d("SmsReceiver", "SMS received from $sender: $content")
-            }
-        }
-    }
-}
